extract fetching describe information from remote class
