+++ lib/cli/fetch.js
@@ -8,6 +8,7 @@ var chalk = require('chalk');
 var Manifest = require('../manifest');
 var CliUtils = require('./utils');
 var config = new(require('../config'))();
+var DescribeRemote = require('../describe-remote');
 
 var doc = "Usage:\n" +
 	"	force-dev-tool fetch [options] [<remote>]\n" +
@@ -51,39 +52,44 @@ SubCommand.prototype.process = function(proc, callback) {
 		var remoteName = client.name;
 		console.log('Fetching from remote ' + chalk.cyan(remoteName));
 		var prefix = '[' + remoteName + ']: ';
-		client.fetch(function(err, result) {
-			if (err) {
-				return remoteCallback(prefix + err);
+		client.login(function(loginErr){
+			if (loginErr) {
+				return remoteCallback(prefix + loginErr);
 			}
-			var manifestDir = path.dirname(self.project.getManifestPath(remoteName));
-			if (!fs.existsSync(manifestDir)) {
-				fs.mkdirSync(manifestDir);
-			}
-			if (result.warnings && result.warnings.length > 0) {
-				console.log(chalk.yellow(result.warnings.join('\n')));
-			}
-			if (result.describeMetadataResult) {
-				var describeMetadataResultPath = self.project.getConfigPath(remoteName) + '-describe-metadata-result.json';
-				fs.writeFileSync(describeMetadataResultPath, JSON.stringify(result.describeMetadataResult));
-				console.log("Created " + chalk.cyan(path.relative(proc.cwd, describeMetadataResultPath)));
-			}
-			if (result.toolingObjects) {
-				var describeToolingObjectsResultPath = self.project.getConfigPath(remoteName) + '-describe-tooling-objects-result.json';
-				fs.writeFileSync(describeToolingObjectsResultPath, JSON.stringify(result.toolingObjects));
-				console.log("Created " + chalk.cyan(path.relative(proc.cwd, describeToolingObjectsResultPath)));
-			}
-			if (result.manifest) {
-				var manifestPath = self.project.getManifestPath(remoteName);
-				fs.writeFile(manifestPath, JSON.stringify(result.manifest), function(writeErr) {
-					if (writeErr) {
-						return remoteCallback(prefix + writeErr);
-					}
-					console.log("Created " + chalk.cyan(path.relative(proc.cwd, manifestPath)));
-					return remoteCallback(null);
-				});
-			} else {
-				return remoteCallback(prefix + "Fetching failed.");
-			}
+			new DescribeRemote(client.conn).fetch(function(err, result) {
+				if (err) {
+					return remoteCallback(prefix + err);
+				}
+				var manifestDir = path.dirname(self.project.getManifestPath(remoteName));
+				if (!fs.existsSync(manifestDir)) {
+					fs.mkdirSync(manifestDir);
+				}
+				if (result.warnings && result.warnings.length > 0) {
+					console.log(chalk.yellow(result.warnings.join('\n')));
+				}
+				if (result.describeMetadataResult) {
+					var describeMetadataResultPath = self.project.getConfigPath(remoteName) + '-describe-metadata-result.json';
+					fs.writeFileSync(describeMetadataResultPath, JSON.stringify(result.describeMetadataResult));
+					console.log("Created " + chalk.cyan(path.relative(proc.cwd, describeMetadataResultPath)));
+				}
+				if (result.toolingObjects) {
+					var describeToolingObjectsResultPath = self.project.getConfigPath(remoteName) + '-describe-tooling-objects-result.json';
+					fs.writeFileSync(describeToolingObjectsResultPath, JSON.stringify(result.toolingObjects));
+					console.log("Created " + chalk.cyan(path.relative(proc.cwd, describeToolingObjectsResultPath)));
+				}
+				if (result.manifest) {
+					var manifestPath = self.project.getManifestPath(remoteName);
+					fs.writeFile(manifestPath, JSON.stringify(result.manifest), function(writeErr) {
+						if (writeErr) {
+							return remoteCallback(prefix + writeErr);
+						}
+						console.log("Created " + chalk.cyan(path.relative(proc.cwd, manifestPath)));
+						return remoteCallback(null);
+					});
+				} else {
+					return remoteCallback(prefix + "Fetching failed.");
+				}
+			});
 		});
 	}, function(err) {
 		if (err) {
